requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
langchain
langchain-openai
langchain-anthropic
openai
openai-whisper
numpy
librosa
soundfile
elevenlabs
httpx
pytest
pypdf
faiss-cpu
langchain-community
chromadb